console: Remove options from device utils
