+++ pkg/webui/console/lib/device-utils.js
@@ -12,12 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import React from 'react'
-
-import Message from '@ttn-lw/lib/components/message'
-
-import sharedMessages from '@ttn-lw/lib/shared-messages'
-
 import randomByteString from '@console/lib/random-bytes'
 
 export const ACTIVATION_MODES = Object.freeze({
@@ -256,38 +250,3 @@ export const lorawanVersionValueSetter = ({ setValues, setFieldTouched }, { valu
     lorawan_phy_version: phyVersions.length === 1 ? phyVersions[0].value : '',
   }))
 }
-
-// 0...7
-export const pingSlotPeriodicityOptions = Array.from({ length: 8 }, (_, index) => {
-  const value = Math.pow(2, index)
-
-  return {
-    value: `PING_EVERY_${value}S`,
-    label: <Message content={sharedMessages.secondInterval} values={{ count: value }} />,
-  }
-})
-// 0...15
-export const adrAckLimitOptions = Array.from({ length: 16 }, (_, index) => {
-  const value = Math.pow(2, index)
-
-  return {
-    value: `ADR_ACK_LIMIT_${value}`,
-    label: <Message content={sharedMessages.adrAckValue} values={{ count: value }} />,
-  }
-})
-// 0...15
-export const adrAckDelayOptions = Array.from({ length: 16 }, (_, index) => {
-  const value = Math.pow(2, index)
-
-  return {
-    value: `ADR_ACK_DELAY_${value}`,
-    label: <Message content={sharedMessages.adrAckValue} values={{ count: value }} />,
-  }
-})
-export const maxDutyCycleOptions = [
-  { value: 'DUTY_CYCLE_1', label: '100%' },
-  { value: 'DUTY_CYCLE_16', label: '6.25%' },
-  { value: 'DUTY_CYCLE_128', label: '0.781%' },
-  { value: 'DUTY_CYCLE_1024', label: '0.098%' },
-  { value: 'DUTY_CYCLE_16384', label: '0.006%' },
-]
